feat: refactor model loading and evaluation scripts
