+++ global_paths.py
@@ -1,10 +1,9 @@
 DATA = "data"
-VIOLATIONS = {
-    'No Violation': 0,
-    'ANSI A13.1 (Pipe Marking)': 1,
-    'ANSI Z358.1-2014 (Emergency Equipment)': 2,
-    'OSHA 1910.157(c)(1)': 3,
-    'OSHA 1910.303(e)(1)': 4,
-    'OSHA 1910.303(g)(1)': 5,
-    'OSHA 1910.37(a)(3)': 6
-}
+VIOLATIONS = [
+    'ANSI A13.1 (Pipe Marking)',
+    'ANSI Z358.1-2014 (Emergency Equipment)',
+    'OSHA 1910.157(c)(1)',
+    'OSHA 1910.303(e)(1)',
+    'OSHA 1910.303(g)(1)',
+    'OSHA 1910.37(a)(3)'
+]
